% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aovtable.R
\name{aovtable}
\alias{aovtable}
\title{Wrapper around \code{\link[stats]{aov}} with better output.}
\usage{
aovtable(formula = NULL, unit = NULL, cv = NULL, data = NULL,
  dbg = FALSE, ...)
}
\arguments{
\item{formula}{RHS of formula should be in partheses, () between any predictor and unit: y~(x1+x2)}

\item{unit}{Factor for stratification units, typicall subject or item, used to build error term}

\item{cv}{Say something here}

\item{data}{Data.frame passed to \code{\link[stats]{aov}}}

\item{dbg}{Enable verbose output for debugging}

\item{...}{Additional arguments passed to \code{\link[stats]{aov}}}
}
\value{
An anova table (data.frame)
}
\description{
Wrapper around \code{\link[stats]{aov}} for repeated measures anova.
}
\details{
No details yet.
}
\author{
David Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link[stats]{aov}}

Other aovtable: \code{\link{print.aovtable}}
}

