% Generated by roxygen2 (4.0.1): do not edit by hand
\name{be.numeric}
\alias{be.numeric}
\title{Checks whether each element in a vector of strings can be interpreted as a number.}
\usage{
be.numeric(c)
}
\arguments{
\item{c}{a vector of strings}
}
\value{
a vector of bool
}
\description{
Checks whether each element in a vector of strings can be interpreted as a number.
}
\details{
Checks whether a (vector of) strings can be interpreted as a number.
}
\author{
cribbed from \url{http://rosettacode.org/wiki/Determine_if_a_string_is_numeric#R}
}

