% Generated by roxygen2 (4.0.1): do not edit by hand
\name{blocks}
\alias{blocks}
\title{Find contiguous blocks of identical integers in a numeric vector.}
\usage{
blocks(v)
}
\arguments{
\item{v:}{a numeric vector containing only whole numbers}
}
\value{
numeric matrix with 3 columns and 1 row for each block.
}
\description{
Find contiguous blocks of identical integers in a numeric vector.
}
\author{
David Braze
}
\seealso{
\code{\link{series}}
}

