% Generated by roxygen2 (4.0.1): do not edit by hand
\name{char2int}
\alias{char2int}
\title{Map character elements to integer elements.}
\usage{
char2int(v)
}
\arguments{
\item{v}{A character vector.}
}
\value{
A vector of class integer.
}
\description{
Map character elements to integer elements.
}
\details{
Maps a character vector to an integer vector. Strings are mapped to integers in the
order that they are encountered, as enumerated from the beginning of the input vector. All
occurances of a given string are mapped to the same integer. Integer enumeration spans
\code{1:n}, where \code{n == length(unique(v))}.

R> v1
 [1] "f" "f" "f" "a" "a" "a" "a" "f" "f" "d" "d" "d" "d" "d" "a" "a"
R> char2int(v1)
f f f a a a a f f d d d d d a a
1 1 1 2 2 2 2 1 1 3 3 3 3 3 2 2
}
\author{
David Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{series}}
}

