% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cocaForm2lemmas}
\alias{cocaForm2lemmas}
\title{Get lemmas for the specified wordform.}
\usage{
cocaForm2lemmas(word, data)
}
\arguments{
\item{word}{A character string holding the wordform}

\item{data}{A data.frame containing CoCA lexical statistics. See \code{\link{cocaReadFreq}}.}
}
\value{
A data.frame with columns w1, L1, WC.
}
\description{
Given a wordform, return all lemmas listed for that form in CoCA, as a 3 column data.frame
containing wordform/lemma/wordclass.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

