% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cocaLemma2forms}
\alias{cocaLemma2forms}
\title{Get attested wordforms for the specified lemma.}
\usage{
cocaLemma2forms(lemma, data)
}
\arguments{
\item{lemma}{A 2 element list with named elements "form" (the lemma form) and "class" (the
simple word class). For acceptable class values see \code{\link{simpleWordClass}}.}

\item{data}{A data.frame containing CoCA lexical statistics. See \code{\link{readCocaFreq}}.}
}
\value{
A character vector containing wordforms.
}
\description{
Given a lemma as a list containing a form/wordclass pair, return all wordforms associated with it
in CoCA.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

