% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cocaLemma2words}
\alias{cocaLemma2words}
\title{Get attested wordforms for the specified lemma.}
\usage{
cocaLemma2words(lemma, data)
}
\arguments{
\item{lemma}{A 2 element character vector. The first element is the lemma form, and the second element is the
simple word class. See \code{\link{cocaSimpleWordClass}} for acceptable word class values.}

\item{data}{A data.frame containing CoCA lexical statistics. See \code{\link{cocaReadFreq}}.}
}
\value{
A character vector containing wordforms.
}
\description{
Given a lemma as a list containing a form/wordclass pair, return all wordforms associated with it
in CoCA.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

