% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cocaLemmaFreq}
\alias{cocaLemmaFreq}
\title{Summed wordform frequencies for a lemma.}
\usage{
cocaLemmaFreq(lemma = "run", WC = NA, data)
}
\arguments{
\item{lemma}{Character string indicating the lemma (form).}

\item{WC}{Character string indicating the wordclass of the lemma.}

\item{data}{A data.frame containing CoCA lexical statistics. See \code{\link{cocaReadFreq}}.}
}
\value{
A data.frame containing columns for lemma (L1), simpleWC (WC), number of simpleWC
included in summary (nWC), number of wordforms included in summary (nforms), summed coca
frequencies for wordforms (coca).
}
\description{
Summed wordform frequencies for a lemma.
}
\details{
Return the summed wordform frequencies for a lemma. If wordclass (WC) is specified,
then the set is restricted to that wordclass.

TODO: Fix content of WC in case of multiple WC (WC=NA). Should be concatenation of actual
values.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

