% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cocaLemmaSet}
\alias{cocaLemmaSet}
\title{Return the set of wordforms associated with a lemma.}
\usage{
cocaLemmaSet(lemma = "run", WC = NA, data)
}
\arguments{
\item{lemma}{Character string indicating the lemma (form).}

\item{WC}{Character string indicating the wordclass of the lemma.}

\item{data}{A data.frame containing CoCA lexical statistics. See \code{\link{cocaReadFreq}}.}
}
\value{
A data.frame containing columns for lemma (L1), simpleWC (WC), CLAWS7 tag (c1), wordform
(w1), coca frequency (coca).
}
\description{
Return the set of wordforms associated with a lemma.
}
\details{
Return the set of wordforms associated with a lemma. If wordclass (WC) is specified,
then the set is restricted to that wordclass.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

