% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coca.R
\name{cocaNeighbors}
\alias{cocaNeighbors}
\title{Find lexical neighbors.}
\usage{
cocaNeighbors(word, data)
}
\arguments{
\item{word}{A character string indicating the word defining the neighborhood.}

\item{data}{A data.frame containing CoCA lexical statistics. See \code{\link{cocaReadFreq}}.}
}
\value{
Returns a data.frame containing 1 row for each neighbor found in \code{data},
specifying: index (ID), wordform (w1), lemma (L1), simple wordclass (WC), CLAWS7 tag (c1), raw
CoCA frequency (coca).
}
\description{
Find lexical neighbors.
}
\details{
Find all neighbors for \code{word} in data (assumes coca data.frame). Neighbors are defined as words
differing from \code{word} by a single letter substitution, deletion or insertion.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

