% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coca.R
\name{cocaWord2lemmas}
\alias{cocaWord2lemmas}
\title{Get lemmas for the specified wordform.}
\usage{
cocaWord2lemmas(word, data)
}
\arguments{
\item{word}{A character string holding the wordform}

\item{data}{A data.frame containing CoCA lexical statistics. See \code{\link{cocaReadFreq}}.}
}
\value{
A character matrix with columns w1 (wordform), L1 (lemma/stem), WC (word class).
}
\description{
Given a wordform, return all lemmas listed for that form in CoCA, as a 3 column data.frame
containing wordform/lemma/wordclass.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

