% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compcorr}
\alias{compcorr}
\title{compcorr()}
\usage{
compcorr(n1, r1, n2, r2)
}
\arguments{
\item{n1}{n for correlation 1}

\item{r1}{r for correlation 1}

\item{n2}{n for correlation 2}

\item{r2}{r for correlation 2}
}
\value{
list containing diff(Z1, Z2) and pval for difference.
}
\description{
Fishers Z for comparing two correllation coefficients.
Method cribbed from:
  http://www.fon.hum.uva.nl/Service/Statistics/Two_Correlations.html
  http://ftp.sas.com/techsup/download/stat/compcorr.html
and implemented in R:
}
\author{
Dave Braze
}

