% Generated by roxygen2 (4.0.1): do not edit by hand
\name{corrTable}
\alias{corrTable}
\title{Get correlation table for numerical matrix M.}
\usage{
corrTable(M, type = "pearson", file = "", diag = FALSE)
}
\arguments{
\item{M}{A numeric matrix with at least 2 columns.}

\item{type}{Passed to \code{\link[Hmisc]{rcor2}}.}

\item{file}{A string. If not empty, then save the matrix to file with the given name.}

\item{diag}{Values to put in diagonal of corrTable: NA (default), standard deviation, variance}
}
\value{
A matrix with R values in the lower triangle and NAs elsewhere
}
\description{
This is wrapper around \code{\link[Hmisc]{rcor2}} that aims to return a reasonably formatted table of
correlations. R values are in the lower triangle with NAs elsewhere. Diagonal is NA by default,
but may be set to standard deviations or variances.
}
\details{
TODO: put means or SDs on the diagonal; put Ns or pvals in the upper.tri
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

