% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosineDist.R
\name{cosinedist}
\alias{cosinedist}
\title{Compute cosine distance.}
\usage{
cosinedist(M)
}
\arguments{
\item{M}{A numeric matrix.}
}
\value{
A matrix with pairwise cosine distances in the lower triangle and vector magnitudes on
the diagonal.
}
\description{
Get cosine distance values for a set of vectors.
}
\details{
Takes a numeric matrix (M) of arbitrary dimension, treats rows of M as vectors and computes the
pairwise cosine distances among the rows of M. Return a matrix of pairwise distances with vector
magnitudes on the diagonal.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{dist}}
}

