% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fact2char}
\alias{fact2char}
\title{fact2char()}
\usage{
fact2char(f)
}
\arguments{
\item{f,}{a factor}
}
\value{
a vector of class "character"
}
\description{
Coerce a factor to character
}
\details{
Preserves level labels associated with each element ....
}
\author{
Dave Braze
}

