% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factors.R
\name{fact2num}
\alias{fact2num}
\title{Coerce a vector of class "factor" to numeric.
Coerces a vector of class factor to numeric. This function preserves underlying integer level
associated with each element of f.}
\usage{
fact2num(f)
}
\arguments{
\item{f}{A vector of class "factor".}
}
\value{
A vector of class "numeric".
}
\description{
Coerce a vector of class "factor" to numeric.
Coerces a vector of class factor to numeric. This function preserves underlying integer level
associated with each element of f.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{fact2char}}
\code{\link{as.numeric}}
}

