% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gammaDistHist}
\alias{gammaDistHist}
\title{Histogram of gamma distribution.}
\usage{
gammaDistHist(x = seq(0, 0.7, by = 0.01), df1 = 6, df2 = 25, mult = 20,
  aspect = 2/5, stack = TRUE)
}
\arguments{
\item{x}{'x' vector argument to \code{\link[stats]{dgamma}}}

\item{df1}{'shape' argument to \code{\link[stats]{dgamma}}}

\item{df2}{'rate' argument to \code{\link[stats]{dgamma}}}

\item{mult}{multiplier used to convert density to counts}

\item{aspect}{aspect ratio of plots; defaults to 2/5}

\item{stack}{stack plots or no; defaults to FALSE (side-by-side plots)}
}
\value{
used for side effect of creating a plot
}
\description{
Histogram of gamma distribution and its reverse.
}
\details{
Histogram of gamma distribution and its reverse.
TODO: use geom_histogram() instead of geom_bar() for consistency with other functions
TODO: label with (mean, median, mode)
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

