% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTrialData}
\alias{getTrialData}
\title{Used by readELascii(). Not intended for end-users.}
\usage{
getTrialData(bounds, lines)
}
\arguments{
\item{bounds}{A numeric tuple. e1 is index marking beginning of trial. e2 is index indicating
end of trial.}

\item{lines}{A vector of strings, each corresponding to 1 line of the EL ASCII file.}
}
\value{
A list of 3 elements, data.frames enumerating fixations, saccades, and blinks for the
trial.
}
\description{
Used by readELascii(). Not intended for end-users. Extract fixations, saccades, and blinks from a trial.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

