% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggQQplot.R
\name{ggQQplot}
\alias{ggQQplot}
\title{Create and plot a QQ plot matrix.}
\usage{
ggQQplot(vars, dat)
}
\arguments{
\item{vars}{A character vector specifying variables with data.frame 'dat' to plot.}

\item{dat}{A data.frame containing variables to be plotted.}
}
\value{
A grid grob created by applying grid.arrange to list of ggplot2 objects.
}
\description{
Given a data.frame and vector of variable names within it, use ggplot2 to create a matrix of quantile-quantile
plots based on the normal distribution and include 'fit' lines on each with intercepts equal to
sample means and slopes equal to sample standard deviations.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{qqp}}
}

