% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4extra.R
\name{glmerProb}
\alias{glmerProb}
\title{Empirical and implied probabilities for binomial outcome models.}
\usage{
glmerProb(model, classes, drop = NA)
}
\arguments{
\item{model}{A glmer object.}

\item{classes}{A character vector containing names of vectors for classification.}

\item{drop}{An integer vector indicating non-categorical terms so that they can be dropped.}
}
\value{
A table consisting of terms built from categorical variables and their linear combinations
}
\details{
Returns a table containing columns:
\enumerate{
    \item First 4 columns as per summary(mod)$coefficients: Estimate, Std. Error, z value, Pr(>|z|)
    \item 'count' number of 1s in response vector
    \item 'n' length of response vector
    \item 'prob.emp' empirical probabilities
    \item 'prob.est.fe' model implied probabilities due to fixed effects only
    \item 'prob.est.re' model implied probabilities for combined fixed and random effects
}

WARNING: this function is very fragile. It is easily broken.
}
\author{
David Braze \email{davebraze@gmail.com}
}

