% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invBoxCox.R
\name{invBoxCox}
\alias{invBoxCox}
\title{Inverse Box-Cox transform.
Compute the inverse Box-Cox transformation of a variable. Requires arguments x and lambda.}
\usage{
invBoxCox(x, lambda)
}
\arguments{
\item{x}{A numerical vector.}

\item{lambda}{Lambda to use for the inverse Box-Cox transform.}
}
\value{
a numerical vector
}
\description{
Inverse Box-Cox transform.
Compute the inverse Box-Cox transformation of a variable. Requires arguments x and lambda.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link[car]{bcPower}}

\code{\link[car]{powerTransform}}
}

