% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.odd.R
\name{is.even}
\alias{is.even}
\title{Is n even?}
\usage{
is.even(n)
}
\arguments{
\item{n}{A numeric vector.}
}
\value{
A logical vector the same length as n. TRUE if n is an integer and is even, FALSE if n
is an integer and not even, NA otherwise.
}
\description{
Return TRUE if n is an integer evenly divisibly by 2.
}
\details{
Test whether elements of a numeric vector are even. For whole numbers return TRUE or FALSE. For decimals return NA.
}
\author{
David Braze \email{davebraze@gmail.com}
}

