% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.numeral.R
\name{is.numeral}
\alias{is.numeral}
\title{Can vector be coerced to numeric?}
\usage{
is.numeral(c)
}
\arguments{
\item{c}{A vector of strings.}
}
\value{
a vector of logicals, NA where is.na(c)
}
\details{
Checks element-wise as to whether each item in a vector of strings can be coerced to numeric.
}
\author{
Basics cribbed from \url{http://rosettacode.org/wiki/Determine_if_a_string_is_numeric#R}

David Braze \email{davebraze@gmail.com}
}

