% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.wholenumber}
\alias{is.wholenumber}
\title{FDB1::is.wholenumber()}
\usage{
is.wholenumber(v, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{v:}{a numeric vector}

\item{tol:}{}
}
\value{
a vector of Logicals
}
\description{
Can x reasonably be interpreted as a whole number?
}
\details{
Takes a numeric vector and returns a Logical vector the same length
}
\author{
Cribbed from R example for base::integer().
}

