% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factors.R
\name{levelOrder}
\alias{levelOrder}
\title{Reorder factor levels.
Reorders the levels of factor f as specified in new.order, which must contain all and only the
existing levels of f, but in the desired order. If new.order and levels(f) do not contain
exactly the same elements, then do nothing.}
\usage{
levelOrder(f, new.order)
}
\arguments{
\item{f}{A factor.}

\item{new.order}{A character vector containing all and only the levels of f, in the desired order.}
}
\value{
A factor.
}
\description{
Reorder factor levels.
Reorders the levels of factor f as specified in new.order, which must contain all and only the
existing levels of f, but in the desired order. If new.order and levels(f) do not contain
exactly the same elements, then do nothing.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{factor}}
\code{\link{levels}}
}

