% Generated by roxygen2 (4.0.1): do not edit by hand
\name{locate}
\alias{locate}
\title{Translate relative coordinates (pos) to data coords.}
\usage{
locate(min, max, pos)
}
\arguments{
\item{min}{minimum bound of plotting region}

\item{max}{maximum bound of plotting region}

\item{pos}{a proportion to be located within min/max}
}
\value{
used for side-effect of creating a plot
}
\description{
Translate relative coordinates (pos) to data coords.
}
\details{
This is a utility function for a translating relative coordinates (pos) to data coordinates. min and max
are values specifiying the full range of a plotting region on the data scale. pos is a
proportion. retval is data-scaled coordinate corresponding to pos.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
See \code{\link[scales]{rescale}} as a possible alternative.
}

