% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mahalDist}
\alias{mahalDist}
\title{Get squared Mahalanobis distances for N observations in M dimensions.}
\usage{
mahalDist(m, scale = TRUE, use = "complete.obs", center = "mean")
}
\arguments{
\item{m}{A data.frame or matrix. Observations in rows.}

\item{scale}{If TRUE, scale variables before getting mahalanobis distances.}

\item{use}{Observations to use in computing covariance matrix. Gets passed to \code{\link[stats]{cov}}.}

\item{center}{Type of univariate center for each variable in \code{m}; "mean" or "median"}
}
\value{
A list with additional class "mahalDist" containing elements:
\enumerate{
    \item{D2: A vector of squared Mahalanobis distances for observations (rows) in \code{m}.
            Incomplete observations return NA.}
    \item{vars: A character vector with the column names from \code{m}.}
    \item{dim: The number of columns of \code{m}.}
}
}
\description{
Get squared Mahalanobis distances for N observations in M dimensions. This is a convenience wrapper
around \code{\link[stats]{mahalanobis}}, which see. Variables are optionally scaled before
distances are computed. Incomplete observations will return NA.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link[stats]{mahalanobis}}

\code{\link[stats]{cov}}
}

