% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matlabCSV}
\alias{matlabCSV}
\title{Function for exporting data.frame to matlab-suitable tab delimited text file.}
\usage{
matlabCSV(DF, name = deparse(substitute(DF)))
}
\arguments{
\item{DF:}{A dataframe to be exported}

\item{name:}{A basename to be used in constructing filenames for key and data files. Defaults to
name of dataframe.}
}
\value{
None. Used for its side effects.
}
\description{
matlabCSV generates 2 ascii files. The first is a tab delimited ascii data  file that contains only
numerical values, or the missing value code 'NaN' (factors in DF are converted to integer
vectors). The second file is a  'key' file that lists all of the variables included (columnwise)
in the data file. For those variables/columns that should be interpreted as factors, gives a list
of the 'levels/labels' that the numerical values in the data file correspond to.
}
\author{
David Braze
}

