% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normal2}
\alias{normal2}
\title{Normal curve plot}
\usage{
normal2(mn1, sd1, mn2, sd2, expanse = 3.5, xlim = NA)
}
\arguments{
\item{mn1}{Mean for curve 1}

\item{sd1}{SD for curve 1}

\item{mn2}{Mean for curve 2}

\item{sd2}{SD for curve 2}

\item{expanse}{multiplier for scaling plot x dimension}

\item{xlim}{'xlim' argument to plot()}
}
\value{
used for side effect of creating a plot.
}
\description{
Two normal curves with (pontentially) different means and SDs.
}
\details{
Plot two normal curves with (pontentially) different means and SDs.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

