% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalDistHist2}
\alias{normalDistHist2}
\title{Histograms for two theoretical normal distributions}
\usage{
normalDistHist2(mn1, sd1, n1, mn2, sd2, n2, stack = FALSE, aspect = 2/5)
}
\arguments{
\item{mn1}{Mean for Distribution 1}

\item{sd1}{SD for Distribution 1}

\item{n1}{N for scaling Histogram 1 (convert density to count)}

\item{mn2}{Mean for Distribution 1}

\item{sd2}{SD for Distribution 1}

\item{n2}{N for scaling Histogram 1 (convert density to count)}

\item{stack}{stack plots vertically or no; defaults to FALSE (side-by-side plots)}

\item{aspect}{aspect ratio for plots; defaults to 2/5}
}
\value{
used for its side-effect of creating a plot
}
\description{
Plot histograms of two different theoretical normal distributions.
}
\details{
Plot histograms of two different theoretical normal distributions, given Mean, SD and N for each.

ToDo:
\enumerate{
    \item print parameters on each facet. See largeNumbers() below.
    \item allow for overlaid histograms (single facet).
    \item add label (geom_text()) for parameters and N to each facet.
    \item add a vlines for means
    \item optionally draw rectangles around mean+-Xsd for each distribution
    \item optionally specify colors of histograms
}
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

