% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalProps}
\alias{normalProps}
\title{Figure showing 68\%-95\%-99.7\% Rule for normal curve.}
\usage{
normalProps(vlines = TRUE, colors = c("gray", "orange", "blue"),
  xlab = "z")
}
\arguments{
\item{vlines}{logical. Use vertical lines to demarcate regions. Default=TRUE}

\item{colors}{character. A vector of 3 color names for each region.}

\item{xlab}{label for horizontal axis. Default = 'z'}
}
\value{
used for side-effect of creating a plot
}
\description{
Figure showing 68\%-95\%-99.7\% Rule for normal curve.
}
\details{
nothing yet
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

