% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalRegion}
\alias{normalRegion}
\title{Plot normal curve with middle region painted.}
\usage{
normalRegion(mn = 100, sd = 10, expanse = 3.5, min, max,
  region.col = "grey", line.col = "black", label = FALSE, lwd = 3)
}
\arguments{
\item{mn}{the mean}

\item{sd}{standard deviation}

\item{expanse}{distance from mean for upper and lower plot limits in sd units. Defaults to 3.5.}

\item{min}{lower bound of painted region}

\item{max}{upper bound of painted region}

\item{region.col}{painted region color}

\item{line.col}{line color}

\item{label}{label for painted region}

\item{lwd}{line width}
}
\value{
Used for side effect.
}
\description{
Plot normal curve with middle region painted.
}
\details{
Note that this handles plots as per function stdnormal1() as a special case.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

