% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributionPlots.R
\name{normalSampHist2}
\alias{normalSampHist2}
\title{Plot histograms for two samples drawn from Normal distributions.}
\usage{
normalSampHist2(mn1, sd1, n1, mn2, sd2, n2, binwidth = mean(c(sd1, sd2))/2,
  stack = FALSE, aspect = 2/5)
}
\arguments{
\item{mn1}{Mean for sample 1}

\item{sd1}{SD for sample 1}

\item{n1}{N for sample 1}

\item{mn2}{Mean for sample 2}

\item{sd2}{SD for sample 2}

\item{n2}{N for sample 2}

\item{binwidth}{binwidth for histogram}

\item{stack}{stack plots vertically or no; defaults to FALSE (side-by-side plots)}

\item{aspect}{set aspect ratio of plot; defaults to 2/5}
}
\value{
used for side-effect of creating a plot
}
\description{
Plot a pair of histograms.
}
\details{
Plot histograms of two samples drawn from different Normal distributions, given Mean, SD and N for each.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

