% Generated by roxygen2 (4.0.1): do not edit by hand
\name{outlierCheck}
\alias{outlierCheck}
\title{Outlier summary.}
\usage{
outlierCheck(x, cut = 2.5)
}
\arguments{
\item{x}{A numerical vector.}

\item{cut}{The cutpoint in sd units; defaults to 2.5.}
}
\value{
A 3-vector with named elements "total", "lower", and "upper" with values indicating the
total number of outliers, outliers in the lower tail and outliers in the upper tail.
}
\description{
Given a numerical vector, indicate number of outliers, and numbers in lower and upper tails.
}
\author{
David Braze \email{davebraze@gmail.com}
}

