% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlierID}
\alias{outlierID}
\title{Locate outliers.}
\usage{
outlierID(x, cut = 2.5)
}
\arguments{
\item{x}{A numerical vector.}

\item{cut}{The cutpoint in sd units; defaults to 2.5.}
}
\value{
A logical vector the length of \code{x}, TRUE wherever \code{x} is more extreme than \code{cut}.
}
\description{
Locate outliers.
}
\details{
Given a numerical vector, locate values of \code{cut} or more standard deviations from the mean.
}
\author{
David Braze \email{davebraze@gmail.com}
}

