% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outliers.R
\name{outlierTrim}
\alias{outlierTrim}
\title{Replace outliers with NA.
Given a numerical vector, replace values of \code{cut} or more standard deviations from the mean with NAs.}
\usage{
outlierTrim(x, cut = 2.5)
}
\arguments{
\item{x}{A numerical vector.}

\item{cut}{The cutpoint in sd units; defaults to 2.5.}
}
\value{
A numerical vector, a copy of \code{x} with outliers replace by NAs.
}
\description{
Replace outliers with NA.
Given a numerical vector, replace values of \code{cut} or more standard deviations from the mean with NAs.
}
\author{
David Braze \email{davebraze@gmail.com}
}

