% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.aovtable.R
\name{print.aovtable}
\alias{print.aovtable}
\title{Prints an anova table generated by aovtable()}
\usage{
\method{print}{aovtable}(x, quote = FALSE, digits = 3, print.mod = FALSE)
}
\arguments{
\item{x}{An aov() model.}

\item{quote}{Print entries with surrounding quotes? Defualts to FALSE.}

\item{digits}{Number of significant digits.}

\item{print.mod}{Print model formula? Defaults to FALSE.}
}
\value{
None. Used for its side effect.
}
\details{
No details yet...
}
\author{
David Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link[stats]{aov}}

\code{\link[base]{print}}

Other aovtable: \code{\link{aovtable}}
}

