% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qqp}
\alias{qqp}
\title{FDB1::qqp()}
\usage{
qqp(v, dat)
}
\arguments{
\item{v}{a character string naming a variable in data.frame 'dat'}

\item{dat}{a data.frame containing at least 1 variable 'v'}
}
\value{
a ggplot2 object
}
\description{
Create a QQ plot for a variable within a data.frame.
}
\details{
Given a data.frame and the name of a variable within it create a quantile-quantile plot based on
the normal distribution. Also plot a 'fit' line with intercept equal to sample mean and slope
equal to sample standard deviation.
}
\author{
Dave Braze
}
\seealso{
\code{\link{qqplot}}
}

