% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qqplot}
\alias{qqplot}
\title{FDB1::qqplot()}
\usage{
qqplot(vars, dat)
}
\arguments{
\item{vars}{a character vector specifying variables with data.frame 'dat' to plot.}

\item{dat}{a data.frame containing variables to be plotted.}
}
\value{
a grid grob created by applying grid.arrange to list of ggplot2 objects
}
\description{
Create and plot a QQ plot matrix
}
\details{
Given a data.frame and vector of variable names within it, create a matrix of quantile-quantile
plots based on the normal distribution and include 'fit' lines on each with intercepts equal to
sample means and slopes equal to sample standard deviations.
}
\author{
Dave Braze
}
\seealso{
\code{\link{qqp}}
}

