% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readELascii}
\alias{readELascii}
\title{Get events from SR Research ASCII data files}
\usage{
readELascii(file = "", tstartre = "TRIALID", tendre = "TRIAL_RESULT",
  eye = NA)
}
\arguments{
\item{file}{string giving path/fname to input file (ELalscii file)}

\item{tstartre}{string containing regular expression that uniquely identifies beginning of trial}

\item{tendre}{string containing regular expression that uniquely identifies end of trial}

\item{eye}{indicates which eye ("R"|"L") to get events from. Currently unused.}
}
\value{
List with one element for the file header and one element for each trial. Each trial
element is itself a list of 3 elements: data.frames enumerating fixations, saccades, and blinks
for the trial.
}
\description{
'readELascii' takes an ASCII data file created by SR Research's EDF2ASC utility and extracts
'eye' events for each trial contained within that file.
}
\details{
SR Research provides a utility (EDF2ASC.exe) that dumps ASCII renderings of their proprietary
EDF data file format. This function reads those ASCII files and extracts eye-movement events
from them (fixations, saccades, blinks).
}
\author{
Dave Braze
}

