% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seP.R
\name{seP}
\alias{seP}
\title{Standard Error of a Proportion.}
\usage{
seP(x)
}
\arguments{
\item{x}{A vector containing only 1s and 0s. NAs are silently dropped.}
}
\value{
The standard error of a proportion.
}
\description{
Standard Error of a Proportion.
}
\author{
David Braze \email{davebraze@gmail.com}
}

