% Generated by roxygen2 (4.0.1): do not edit by hand
\name{series}
\alias{series}
\title{Find sequences of positive integers in numeric vector v.}
\usage{
series(v)
}
\arguments{
\item{v}{a numeric vector.}
}
\value{
numeric matrix with 3 columns and 1 row for each series.
}
\description{
Find sequences of positive integers in numeric vector v.
}
\author{
David Braze
}
\seealso{
\code{\link{blocks}}
}

