% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aovtable.R
\name{aovtable}
\alias{aovtable}
\title{Wrapper around \code{\link[stats]{aov}} with better output.}
\usage{
aovtable(formula = NULL, unit = NULL, cv = NULL, data = NULL,
  dbg = FALSE, ...)
}
\arguments{
\item{formula}{: rhs of formula should be in partheses, () between any predictor and unit: y~(x1+x2)}

\item{unit}{: factor for stratification units, typicall subject or item, used to build error term}

\item{cv}{: say something here}

\item{data}{: data.frame passed to \code{\link[stats]{aov}}}

\item{dbg}{: enable verbose output for debugging}

\item{...}{: additional arguments passed to \code{\link[stats]{aov}}}
}
\value{
an anova table (data.frame)
}
\description{
Wrapper around \code{\link[stats]{aov}} for repeated measures anova.
}
\author{
David Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link[stats]{aov}}

Other aovtable: \code{\link{print.aovtable}}
}

