% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aovtable}
\alias{aovtable}
\title{Wrapper around aov() for repeated measures anova.}
\usage{
aovtable(formula = NULL, unit = NULL, cv = NULL, data = NULL,
  dbg = FALSE, ...)
}
\arguments{
\item{formula}{rhs of formula should be in partheses, () between any predictor and unit: y~(x1+x2)}

\item{unit}{factor for stratification units, typicall subject or item, used to build error term}

\item{cv}{say something here}

\item{data}{data.frame passed to aov()}

\item{dbg}{enable verbose output for debugging}

\item{...}{additional arguments passed to aov()}
}
\value{
an anova table (data.frame)
}
\description{
No details yet...
}
\author{
David Braze
}

