% Generated by roxygen2 (4.0.2): do not edit by hand
\name{be.numeric}
\alias{be.numeric}
\title{Checks element-wise if character vector can be coerced to numeric.}
\usage{
be.numeric(c)
}
\arguments{
\item{c}{: a vector of strings}
}
\value{
a vector of logicals
}
\description{
Checks whether each element in a vector of strings can be coerced to numeric.
}
\author{
cribbed from \url{http://rosettacode.org/wiki/Determine_if_a_string_is_numeric#R}
}

