% Generated by roxygen2 (4.0.1): do not edit by hand
\name{betas}
\alias{betas}
\title{Extract betas from an lm object.}
\usage{
betas(lm)
}
\arguments{
\item{lm}{: an lm object}
}
\value{
A vector of standardized coefficients.
}
\description{
There is no built in method for extracting standardized regression coefficients
(betas) from an \code{\link[stats]{lm}} object. This function fills that gap.
}
\author{
David Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link[stats]{lm}}
}

