% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/betas.R
\name{betas}
\alias{betas}
\title{Extract betas from an lm object.}
\usage{
betas(lm)
}
\arguments{
\item{lm}{An lm object.}
}
\value{
A vector of standardized coefficients.
}
\description{
Extract betas from an lm object.
}
\details{
There is no built in method for extracting standardized regression coefficients (betas)
from an \code{\link[stats]{lm}} object. This function fills that gap.
}
\author{
David Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link[stats]{lm}}
}

