% Generated by roxygen2 (4.0.2): do not edit by hand
\name{blockRandomLists}
\alias{blockRandomLists}
\title{Block randomize a set of trials.}
\usage{
blockRandomLists(nLists = 4, repBlock = 2, nCond, data)
}
\arguments{
\item{nLists}{Number of randomized lists to create (default=4).}

\item{repBlock}{Number of repetitions/condition within each block (default=2).}

\item{nCond}{Number of conditions in data.}

\item{data}{A data.frame containing specification of all conditions/items, sorted by condition.}
}
\value{
A list of data.frames
}
\description{
Block randomize a set of trials.
}
\details{
Generate a specified number of uniquely block randomized copies of a stimulus list.

TODO: this function assumes an equal number of items per condition. extend it to work with a
specified number of items per condition.
}
\author{
David Braze \email{davebraze@gmail.com}
}

