% Generated by roxygen2 (4.0.1): do not edit by hand
\name{blocks}
\alias{blocks}
\title{FDB1::blocks()}
\usage{
blocks(v)
}
\arguments{
\item{v:}{a numeric vector containing only whole numbers}
}
\value{
numeric matrix with 3 columns and 1 row for each block.
}
\description{
FDB1::blocks()
}
\author{
David Braze
}

