% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cocaReadFreq}
\alias{cocaReadFreq}
\title{Read CoCA word frequency table.}
\usage{
cocaReadFreq(file, sep = "\\t", na.strings = "  ",
  quote = "\\"", header = TRUE, simpleWC = TRUE, ...)
}
\arguments{
\item{file}{per \code{\link{read.table}}}

\item{sep}{per \code{\link{read.table}}}

\item{na.strings}{per \code{\link{read.table}}}

\item{quote}{per \code{\link{read.table}}}

\item{header}{per \code{\link{read.table}}}

\item{simpleWC}{If TRUE then add vector of simplified wordclasses to data.frame. See
\code{\link{simpleWordClass}}.}

\item{...}{additional arguments will be passed to \code{\link{read.table}}.}
}
\value{
a data.frame
}
\description{
A convenience wrapper around \code{\link{read.table}} with reasonable defaults for reading the
Corpus of Contemporary American English word frequency file (\url{corpus.byu.edu}). The file
contains tab delimited text, with some idiosynchracies.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{read.table}}
}

