% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cocaSimpleWordClass}
\alias{cocaSimpleWordClass}
\title{Simplify CoCA (CLAWS7) part-of-speech tags.}
\usage{
cocaSimpleWordClass(wordclass)
}
\arguments{
\item{wordclass}{One of the CLAWS7 tags used in CoCA}
}
\value{
A string containing the simplified wordclass tag.
}
\description{
CoCA uses a subset of wordclass tags from the CLAWS7 tagset
(http://ucrel.lancs.ac.uk/claws7tags.html). This function reduces those to a simplified set
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

