% Generated by roxygen2 (4.0.1): do not edit by hand
\name{corrTable}
\alias{corrTable}
\title{Get correlation table for numerical matrix M.}
\usage{
corrTable(M, type, file = "")
}
\arguments{
\item{M}{: a numeric matrix with at least 2 columns}

\item{type}{: passed to Hmisc::rcor2}

\item{file}{: a string. If not empty, then save the matrix to file with the given name.}
}
\value{
A matrix with R values in the lower triangle and NAs elsewhere
}
\description{
Get correlation table for numerical matrix M.
}
\details{
This is wrapper around \code{[Hmisc]{rcor2}} that aims to return a reasonably formatted table of
correlations. R values are in the lower triangle with NAs elsewhere, including the diagonal.

TODO: put means or SDs on the diagonal; put Ns or pvals in the upper.tri
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

