% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fact2char}
\alias{fact2char}
\title{Coerce a vector of class "factor" to character}
\usage{
fact2char(f)
}
\arguments{
\item{f}{A vector of class factor.}
}
\value{
A vector of class "character".
}
\description{
Coerce a vector of class "factor" to character
}
\details{
Coerces a vector of class factor to numeric. This function preserves level labels associated
with each element of f.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{fact2num}}
}

