% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fact2num}
\alias{fact2num}
\title{Coerce a vector of class "factor" to numeric}
\usage{
fact2num(f)
}
\arguments{
\item{f}{a vector of class "factor"}
}
\value{
a vector of class "numeric"
}
\description{
Coerce a vector of class "factor" to numeric
}
\details{
Coerces a vector of class factor to numeric. This function preserves underlying integer level
associated with each element of f.
}
\author{
Dave Braze
}
\seealso{
\code{\link{fact2char}}
}

