% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fact2num}
\alias{fact2num}
\title{fact2num()}
\usage{
fact2num(f)
}
\arguments{
\item{f,}{a factor}
}
\value{
a vector of class "numeric"
}
\description{
Coerce a factor to numeric
}
\details{
Preserves underlying integer level associated with each element ....
}
\author{
Dave Braze
}

