% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getTrialData}
\alias{getTrialData}
\title{FDB1::getEyeEvents()}
\usage{
getTrialData(bounds, lines)
}
\arguments{
\item{bounds}{A numeric tuple. e1 is index marking beginning of trial. e2 is index indicating
end of trial.}

\item{lines}{A vector of strings, each corresponding to 1 line of the EL ASCII file.}
}
\value{
List with one element for the file header and one element for each trial. Each trial
element is itself a list of 3 elements: data.frames enumerating fixations, saccades, and blinks
for the trial.
}
\description{
Used by readELascii(). Not intended for end-users.
}
\details{
Used by readELascii(). Not intended for end-users.
}
\author{
Dave Braze
}

