% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lme4extra.R
\name{ggCaterpillar}
\alias{ggCaterpillar}
\title{Plot random effects from lmer fit.}
\usage{
ggCaterpillar(re, QQ = TRUE, likeDotplot = TRUE)
}
\arguments{
\item{re}{An object of class ranef.mer.}

\item{QQ}{If TRUE, create QQ plot. If FALSE create caterpillar plot.}

\item{likeDotplot}{Imitate dotplot() -> same scales for random effects.}
}
\value{
A \code{\link{ggplot2}} object.
}
\description{
Plot random effects from lmer fit.
}
\details{
Create ggplot based QQ plot or caterpillar plot for \code{\link{ranef}} from \code{\link{mer}}
object.
}
\author{
\url{http://stackoverflow.com/users/484139/caracal}
\url{http://stackoverflow.com/users/1857266/didzis-elferts}
}
\seealso{
Original code cribbed from here:
\url{http://stackoverflow.com/questions/13847936/in-r-plotting-random-effects-from-lmer-lme4-package-using-qqmath-or-dotplot}
}

