% Generated by roxygen2 (4.0.1): do not edit by hand
\name{invBoxCox}
\alias{invBoxCox}
\title{Inverse Box-Cox transform}
\usage{
invBoxCox(x, lambda)
}
\arguments{
\item{x}{: a numerical vector}

\item{lambda}{: lambda to use for the inverse Box-Cox transform}
}
\value{
a numerical vector
}
\description{
Inverse Box-Cox transform.
}
\details{
Compute the inverse Box-Cox transformation of a variable. Requires arguments x and lambda.
}
\author{
Dave Braze
}
\seealso{
\code{\link{car::bcPower}}

\code{\link{car::powerTransform}}
}

