% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplot.R
\name{iplot}
\alias{iplot}
\title{Plot eye-movement summaries by sentence region.}
\usage{
iplot(x.factor, trace.factor, response, fun = function(x) {     mean(x, na.rm
  = TRUE) }, type = c("l", "p", "b", "o"), legend = TRUE,
  leg.title = deparse(substitute(trace.factor)), trace.labels = NULL,
  fixed = FALSE, xlab = deparse(substitute(x.factor)), ylab = ylabel,
  main = "", ylim = range(cells, na.rm = TRUE), lty = nc:1, lwd = 1,
  col = 1, pch = c(1:9, 0, letters), xpd = NULL, leg.bg = par("bg"),
  leg.bty = "n", point.cex = 1, leg.x = 1, leg.y = ylim[2],
  xtick = FALSE, xaxt = par("xaxt"), axes = TRUE, ...)
}
\arguments{
\item{x.factor}{Factor for x axis, usually sentence region.}

\item{trace.factor}{Factor for separate traces, usually sentence condition.}

\item{response}{Response measure to be plotted on y axis.}

\item{fun}{Function for summarizing response measure, defaults to mean(x, na.rm=T).}

\item{type}{This and following params are as per plot() and par(), which see.}

\item{...}{other parameters that will be passed to matplot()}
}
\value{
None. Used for its side effects.
}
\details{
iplot() is based on interaction.plot(). Be careful of missing values in "response"; make sure
that "fun" does the right thing with them.
}
\author{
David Braze \email{davebraze@gmail.com}
}

