% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.numeral.R
\name{is.numeral}
\alias{is.numeral}
\title{Checks element-wise if character vector can be coerced to numeric.}
\usage{
is.numeral(c)
}
\arguments{
\item{c}{: a vector of strings}
}
\value{
a vector of logicals, NA where is.na(c)
}
\description{
Checks whether each element in a vector of strings can be coerced to numeric.
}
\author{
Basics cribbed from \url{http://rosettacode.org/wiki/Determine_if_a_string_is_numeric#R}

David Braze \email{davebraze@gmail.com}
}

