% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.odd}
\alias{is.odd}
\title{FDB1::is.odd()}
\usage{
is.odd(v)
}
\arguments{
\item{v}{a numeric vector}
}
\value{
TRUE if n is an integer and is odd, FALSE if n is an integer and not odd, NA otherwise.
}
\description{
Returns TRUE if n is odd.
}
\author{
David Braze
}

