% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.odd}
\alias{is.odd}
\title{Returns TRUE where n is odd.}
\usage{
is.odd(n)
}
\arguments{
\item{n}{: a numeric vector}
}
\value{
TRUE if n is an integer and is odd, FALSE if n is an integer and not odd, NA otherwise.
}
\description{
Returns TRUE where n is odd.
Test whether elements of numeric vector are odd. For whole numbers return TRUE or FALSE. For decimals return NA.
}
\author{
David Braze \email{davebraze@gmail.com}
}

