% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Can each element of x reasonably be interpreted as a whole number?}
\usage{
is.wholenumber(v, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{v}{: a numeric vector}

\item{tol}{: what counts as close enough to 0? Defaults to .Machine$double.eps^0.5}
}
\value{
a vector of Logicals
}
\description{
Takes a numeric vector and returns a Logical vector the same length
}
\author{
Cribbed from example in \code{\link[base]{integer}}
}

