% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Can each element of x reasonably be interpreted as a whole number?
Takes a numeric vector and returns a Logical vector the same length}
\usage{
is.wholenumber(v, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{v}{A numeric vector.}

\item{tol}{What counts as close enough to 0? Defaults to .Machine$double.eps^0.5.}
}
\value{
A vector of Logicals.
}
\description{
Can each element of x reasonably be interpreted as a whole number?
Takes a numeric vector and returns a Logical vector the same length
}
\author{
Cribbed from example in \code{\link[base]{integer}}
}

