% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mahalDist.R
\name{mahalPlot}
\alias{mahalPlot}
\title{QQplot of squared mahalanobis distance against Chi-square distribution.
QQplot of squared mahalanobis distance against Chi-square distribution.}
\usage{
mahalPlot(X, envp = 0.95, col.point = "blue", ...)
}
\arguments{
\item{envp}{Width of confidence envelope passed to \code{\link[car]{qqPlot}}. Defaults to .95.}

\item{D2}{Vector of squared mahalanobis distances as calculated, for example, by \code{\link{mahalDist}}.}

\item{dim}{Number of variables (dimensions) entering into D2.}

\item{point.col}{Color for plotting points passed to \code{\link[car]{qqPlot}}. Defaults to "blue".}
}
\value{
Used for the side effect of creating a plot.
}
\description{
QQplot of squared mahalanobis distance against Chi-square distribution.
QQplot of squared mahalanobis distance against Chi-square distribution.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link[stats]{mahalanobis}}

\code{\link{mahalDist}}

\code{\link[car]{qqPlot}}
}

