% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nobs}
\alias{nobs}
\title{Number of non-missing values.}
\usage{
nobs(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
integer, length of x minus the number of NAs in x.
}
\description{
Returns the number of non-missing values in vector x. A convenience wrapper around sum(!is.na(x)).
}
\author{
David Braze \email{davebraze@gmail.com}
}

