% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nobs}
\alias{nobs}
\title{A convenience wrapper around sum(!is.na(x)).}
\usage{
nobs(v)
}
\arguments{
\item{v}{a vector}
}
\value{
integer, length of x minus the number of NAs in x.
}
\description{
nobs() returns the number of non-missing values in vector x.
}
\author{
David Braze
}

