% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qqp}
\alias{qqp}
\title{Create a QQ plot for a variable within a data.frame.}
\usage{
qqp(v, dat)
}
\arguments{
\item{v}{: a character string naming a variable in data.frame 'dat'}

\item{dat}{: a data.frame containing at least 1 variable 'v'}
}
\value{
a \pkg{ggplot2} object
}
\description{
Given a data.frame and the name of a variable within it create a quantile-quantile plot based on
the normal distribution. Also plot a 'fit' line with intercept equal to sample mean and slope
equal to sample standard deviation.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{qqplot}}
}

