% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggQQplot.R
\name{qqp}
\alias{qqp}
\title{Quantile-Quantile plot for a variable within a data.frame.}
\usage{
qqp(v, dat)
}
\arguments{
\item{v}{A character string naming a variable in data.frame 'dat'.}

\item{dat}{A data.frame containing at least 1 variable 'v'.}
}
\value{
a \pkg{ggplot2} object
}
\description{
Quantile-Quantile plot for a variable within a data.frame.
}
\details{
Given a data.frame and the name of a variable within it use ggplot2 to create a
quantile-quantile plot based on the normal distribution. Also plot a 'fit' line with intercept
equal to sample mean and slope equal to sample standard deviation.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{ggQQplot}}
}

