% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regtable.R
\name{regtable}
\alias{regtable}
\title{Customized regression table.}
\usage{
regtable(lmobj, intercept = FALSE, ...)
}
\value{
A formatted regression table.
}
\description{
Customized regression table.
}
\details{
No details yet...
}
\author{
David Braze \email{davebraze@gmail.com}
}

