% Generated by roxygen2 (4.0.1): do not edit by hand
\name{se}
\alias{se}
\title{Standard error of the mean.}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{Object to compute SEMs for. Can be vector, matrix or data.frame.}

\item{na.rm}{Specify how to handle missing values.}
}
\value{
Standard error of the mean for x, or each column of x.
}
\description{
Standard error of the mean.
}
\author{
David Braze \email{davebraze@gmail.com}
}

