% Generated by roxygen2 (4.0.1): do not edit by hand
\name{series}
\alias{series}
\title{Find sequences of integers in numeric vector v.}
\usage{
series(v, step = 1, minseries = 2)
}
\arguments{
\item{v}{: a numeric vector containing only whole numbers.}

\item{step}{: The step size and direction (distance between adjacent items) for identifying
series. Note that the sign of the step matters. Use positive values for monotonically increasing
series and negative values for monotonically decreasing series. Defaults to 1.}

\item{minseries}{: The minimimum series length. Defaults to 2.}
}
\value{
numeric matrix with 3 columns and 1 row for each series.
}
\description{
Find sequences of integers in numeric vector v.
}
\details{
Takes a vector of whole numbers and returns a matrix with 3 columns: (1) integer at
beginning of series, (2) index pointing to beginning of series, (3) length of series
}
\author{
David Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{blocks}}
}

